backend/requirements.txt
fastapi
uvicorn
requests
openai-whisper
sqlalchemy
passlib[argon2,bcrypt]
pydantic
python-jose[cryptography]
python-multipart
pyinstaller
aiofiles
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
psycopg2-binary
alembic
numpy
librosa
